+++ web/src/_interfaces/IMonthAndYear.ts
@@ -1,0 +1,4 @@
+export interface IMonthAndYear {
+  month: number
+  year: number
+}