+++ history.md
@@ -1,10 +1,11 @@
 9.25
 
 구현
-modal 입력 받아서 db 저장
 
+- modal 입력 받아서 db 저장
 - 시간 입력 이상할떄 에러 처리
-  db request Data해서 display
+- 이미 일정이 있을 때 에러 처리
 
 디버깅
+db request Data해서 display
 모달에서 한글 입력 잘 안됨.. keydown 을 바꿔야 할듯